search view created, sorting project's folders
