+++ app/src/main/java/com/globant/imdb/view/home/PlayFragment.kt
@@ -1,4 +1,4 @@
-package com.globant.imdb.home
+package com.globant.imdb.view.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.globant.imdb.R
 
-class SearchFragment : Fragment() {
+class PlayFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_search, container, false)
+        return inflater.inflate(R.layout.fragment_play, container, false)
     }
 }
