+++ app/src/main/java/com/globant/imdb/view/home/SearchFragment.kt
@@ -1,0 +1,28 @@
+package com.globant.imdb.view.home
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.globant.imdb.databinding.FragmentSearchBinding
+
+class SearchFragment : Fragment() {
+
+    private var binding: FragmentSearchBinding? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        // Inflate the layout for this fragment
+        val fragmentBinding = FragmentSearchBinding.inflate(inflater, container, false)
+        binding = fragmentBinding
+        return fragmentBinding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        binding = null
+    }
+}
