+++ app/src/main/java/com/globant/imdb/view/login/LoginFragment.kt
@@ -1,7 +1,6 @@
-package com.globant.imdb
+package com.globant.imdb.view.login
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
