+++ app/src/main/java/com/globant/imdb/view/registration/RegistrationFragment.kt
@@ -1,13 +1,13 @@
-package com.globant.imdb
+package com.globant.imdb.view.registration
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.content.res.ResourcesCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.globant.imdb.R
 import com.globant.imdb.databinding.FragmentRegistrationBinding
 
 class RegistrationFragment : Fragment() {
@@ -27,7 +27,9 @@ class RegistrationFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         binding?.apply {
             registrationFragment = this@RegistrationFragment
-            usernameFrame.typeface = context?.let { ResourcesCompat.getFont(it, R.font.roboto_light_italic) }
+            usernameFrame.typeface = context?.let { ResourcesCompat.getFont(it,
+                R.font.roboto_light_italic
+            ) }
             usernameInput.typeface = context?.let { ResourcesCompat.getFont(it, R.font.roboto) }
         }
     }
