+++ app/src/main/java/com/globant/imdb/view/splash/SplashFragment.kt
@@ -1,6 +1,5 @@
-package com.globant.imdb
+package com.globant.imdb.view.splash
 
-import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
 import androidx.fragment.app.Fragment
@@ -8,6 +7,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
+import com.globant.imdb.R
 
 class SplashFragment : Fragment() {
