+++ app/src/main/res/layout/fragment_bottom_nav.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".home.BottomNavFragment">
+        tools:context=".view.home.BottomNavFragment">
 
         <FrameLayout
             android:id="@+id/bottomNav_frame_layout"
