+++ app/src/main/res/layout/fragment_play.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.PlayFragment">
+    tools:context=".view.home.PlayFragment">
 
     <TextView
         android:id="@+id/textView2"
