+++ app/src/main/res/layout/fragment_registration.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".view.registration.RegistrationFragment">
 
     <data>
         <variable
             name="registrationFragment"
-            type="com.globant.imdb.RegistrationFragment" />
+            type="com.globant.imdb.view.registration.RegistrationFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".RegistrationFragment">
+        android:layout_height="match_parent">
 
         <ImageView
             android:id="@+id/imdb_register_logo"
