+++ app/src/main/res/layout/fragment_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/imdb_yellow"
-    tools:context=".SplashFragment">
+    tools:context=".view.splash.SplashFragment">
 
     <TextView
         android:id="@+id/imdb_logo"
