+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/splashFragment"
-        android:name="com.globant.imdb.SplashFragment"
+        android:name="com.globant.imdb.view.splash.SplashFragment"
         android:label="fragment_splash"
         tools:layout="@layout/fragment_splash" >
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.globant.imdb.LoginFragment"
+        android:name="com.globant.imdb.view.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -25,7 +25,7 @@
     </fragment>
     <fragment
         android:id="@+id/registrationFragment"
-        android:name="com.globant.imdb.RegistrationFragment"
+        android:name="com.globant.imdb.view.registration.RegistrationFragment"
         android:label="fragment_registration"
         tools:layout="@layout/fragment_registration" >
         <action
@@ -34,7 +34,7 @@
     </fragment>
     <fragment
         android:id="@+id/bottomNavFragment"
-        android:name="com.globant.imdb.home.BottomNavFragment"
+        android:name="com.globant.imdb.view.home.BottomNavFragment"
         android:label="fragment_bottomNav"
         tools:layout="@layout/fragment_bottom_nav" />
 </navigation>
